adapted some aspects

+++ src/parkinglot/Vehicle.java
@@ -1,5 +1,7 @@
 package parkinglot;
 
+import java.math.BigDecimal;
+
 /*
  * Copyright (c) 2019. Schenker AG
  * All rights reserved.
@@ -9,21 +11,21 @@ public class Vehicle {
 
     private final int id;
     private final int spacesNeeded;
-    private final double costFactor;
+    private final BigDecimal costFactor;
     private String vehicleType;
 
-    public Vehicle(String vehicleType, int id, int spacesNeeded, double costFactor) {
+    public Vehicle(String vehicleType, int id, int spacesNeeded, BigDecimal costFactor) {
         this.vehicleType = vehicleType;
         this.id = id;
         this.spacesNeeded = spacesNeeded;
-        this.costFactor = costFactor; // setters instead
+        this.costFactor = costFactor;
     }
 
     public int getSpacesNeeded() {
         return spacesNeeded;
     }
 
-    public double getCostFactor() {
+    public BigDecimal getCostFactor() {
         return costFactor;
     }
